+++ packages/datagraphic/stories/DataGraphic.stories.js
@@ -1,7 +1,12 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { withKnobs } from "@storybook/addon-knobs";
+import Borders from "./Borders.svelte";
 
 export default {
   title: "DataGraphic",
   decorators: [withKnobs],
 };
+
+export const borders = () => ({
+  Component: Borders,
+});