+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Change `xDomainMin` / `xDomainMax` & y equivalents to `xMin` / `xMax` etc. ([#95](https://github.com/graph-paper-org/graph-paper/pull/95))
 - Refactor width / height of `DataGraphic` component to reactively update ([#83](https://github.com/graph-paper-org/graph-paper/pull/82))
+- Change `borderThickness` to `borderSize` in `DataGraphic` ([#96](https://github.com/graph-paper-org/graph-paper/pull/96))
 
 ## 0.0.0-alpha.8